auto

+++ src/app/page/auto/fileDetails.tsx
@@ -5,13 +5,12 @@ import { Button, Card, Col, Collapse, Row, Space, Spin, Typography } from 'antd'
 import { AppDispatch, AppState } from 'app/model'
 import IonIcon from 'shared/antd/ionicon'
 import AccountInfo from './accountInfo'
-import {
-  deleteRecipient,
-  removeRecipients,
-} from 'app/model/recipients.controller'
+import { deleteRecipient } from 'app/model/recipients.controller'
 import InputInfoTransfer from 'app/components/inputInfoTransfer'
 import { CollapseAddNew } from 'app/constants'
 import { WrapTotal } from 'app/components/cardTotal'
+import AccountInfoHeader from './accountInfoHeader'
+import { isEqual } from 'app/helper/equals'
 
 const ActionButton = ({
   activeKey = '',
@@ -54,10 +53,11 @@ const ActionButton = ({
   )
 }
 
-const FileDetails = () => {
+const FileDetails = ({ onRemove }: { onRemove: () => void }) => {
   const dispatch = useDispatch<AppDispatch>()
   const {
     recipients: { recipients },
+    main: { fileName },
   } = useSelector((state: AppState) => state)
   const [selected, setSelected] = useState(false)
   const [activeKey, setActiveKey] = useState<string>()
@@ -73,15 +73,36 @@ const FileDetails = () => {
     }
     return setWalletsSeletectd(nextWalletSelected)
   }
+  const onSelectAll = (checked: boolean) => {
+    if (!recipients) return
+    const recipientAddresses = Object.keys(recipients)
+    let nextWalletSelected = [...recipientAddresses]
+    if (!checked) nextWalletSelected = []
+    return setWalletsSeletectd(nextWalletSelected)
+  }
   const onDelete = async () => {
     if (!walletsSelected.length || !Object.keys(recipients).length) return
     setLoading(true)
-    await walletsSelected.forEach((walletAddress) => {
-      if (!!recipients[walletAddress])
-        return dispatch(deleteRecipient({ walletAddress }))
-    })
+    if (isEqual(walletsSelected, Object.keys(recipients))) onRemove()
+    else
+      await walletsSelected.forEach((walletAddress) => {
+        if (!!recipients[walletAddress])
+          return dispatch(deleteRecipient({ walletAddress }))
+      })
     return setLoading(false)
   }
+
+  const errorData = Object.keys(recipients).filter((addr) => {
+    const { amount, email } = recipients[addr]
+    return !amount || !email
+  })
+
+  const listRecipients = Object.keys(recipients).filter((addr) => {
+    const { amount, email } = recipients[addr]
+    return amount && email
+  })
+
+  const existErrorData = !!errorData.length
 
   return (
     <Row gutter={[16, 16]}>
@@ -92,7 +113,7 @@ const FileDetails = () => {
               <Space>
                 <IonIcon name="document-attach-outline" />
                 <Typography.Text style={{ color: 'inherit' }}>
-                  {'File name.csv'}
+                  {fileName}
                 </Typography.Text>
               </Space>
             </Col>
@@ -102,7 +123,7 @@ const FileDetails = () => {
                 size="small"
                 style={{ color: 'inherit' }}
                 icon={<IonIcon name="close-outline" />}
-                onClick={() => dispatch(removeRecipients())}
+                onClick={onRemove}
               />
             </Col>
           </Row>
@@ -146,14 +167,42 @@ const FileDetails = () => {
             <Spin spinning={loading}>
               <Card bordered={false} className="card-content">
                 <Row gutter={[8, 8]}>
-                  {recipients &&
-                    Object.keys(recipients).map((addr, idx) => (
+                  <Col span={24}>
+                    <AccountInfoHeader
+                      selected={selected}
+                      error={existErrorData}
+                      onChecked={onSelectAll}
+                    />
+                  </Col>
+                  {errorData &&
+                    errorData.map((addr, idx) => (
+                      <Col
+                        className={
+                          idx + 1 === errorData.length
+                            ? 'last-item-error-data'
+                            : ''
+                        }
+                        span={24}
+                        key={addr + idx}
+                      >
+                        <AccountInfo
+                          accountAddress={addr}
+                          selected={selected}
+                          onChecked={onSelected}
+                          error={existErrorData}
+                          walletsSelected={walletsSelected}
+                        />
+                      </Col>
+                    ))}
+                  {listRecipients &&
+                    listRecipients.map((addr, idx) => (
                       <Col span={24} key={addr + idx}>
                         <AccountInfo
                           accountAddress={addr}
-                          index={idx}
                           selected={selected}
                           onChecked={onSelected}
+                          error={existErrorData}
+                          walletsSelected={walletsSelected}
                         />
                       </Col>
                     ))}
